Created Readme File For The Project
The Art With Machine Learning Project takes a picture from the camera feed and renders the picture in the styles of various artists around the world. The program has the capability to learn new artistic styles through the use of Tensorflow and the Nvidia CUDA/cudNN libraries to speed up the training process of new styles through the use of GPU Hardware Acceleration.